API/requirements.txt
absl-py==1.4.0
anyio==3.6.2
astunparse==1.6.3
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
ecdsa==0.18.0
fastapi==0.89.1
flatbuffers==23.1.21
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.1
h11==0.14.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.0.0
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
passlib==1.7.4
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
PyMySQL==1.0.2
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.0
starlette==0.22.0
termcolor==2.2.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.12.0
